Fix autoload layout check to allow vendor path naming
This fixes a problem introduced by b91d0058da6265ade231bfcd5516bb0050ecc3d5
if you have a module checked out in a directory which is actually the module
name you won't have a problem.

But when the directory is named "vendor-module" or "puppet-module" the check
fails.

The named commit added a slash as prefix to avoid detecting wrong module names,
I've added a dash.